clean and consolidate EDA work in progress
